pyproject.toml
[project]
name = "prj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "hydra-core>=1.3.2",
    "imageio>=2.37.2",
    "matplotlib>=3.10.7",
    "omegaconf>=2.3.0",
    "pillow>=12.0.0",
    "pycryptodome>=3.23.0",
    "tinyec>=0.4.0",
]
